refactor: WhitePieces와 BlackPieces의 중복 제거를 위해 Pieces로 통합
