--- src/test/java/chess/domain/board/ChessBoardTest.java
@@ -1,13 +1,0 @@
-package chess.domain.board;
-
-/**
- *
- *    @author AnHyungJu, LeeHoBin
- */
-public class ChessBoardTest {
-	// @DisplayName("체스보드 초기화가 정상적으로 되는지 확인")
-	// @Test
-	// void test1() {
-	// 	assertThat(ChessBoard.initialize()).isInstanceOf(ChessBoard.class);
-	// }
-}
