+++ src/test/java/chess/domain/piece/WhitePiecesFactoryTest.java
@@ -14,6 +14,6 @@ class WhitePiecesFactoryTest {
 	@Test
 	void createTest() {
 		assertThat(WhitePiecesFactory.create())
-			.isInstanceOf(WhitePieces.class);
+			.isInstanceOf(Pieces.class);
 	}
 }
